requirements.txt
biopython==1.85
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
Flask==3.1.2
google-auth==2.49.0
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pillow==11.3.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.5
requests==2.32.5
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
Werkzeug==3.1.4
zipp==3.23.0
